backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
redis[hiredis]==5.2.1
httpx==0.28.1
anthropic==0.42.0
qdrant-client==1.12.1
numpy==2.2.1

# Rate Limiting
slowapi==0.1.9
